+++ src/commands/apply.js
@@ -30,6 +30,7 @@ ApplyCommand.args = [
 ]
 
 ApplyCommand.flags = {
+  init: flags.boolean({ char: 'i', description: 'create non-existent configuration files' }),
   keep: flags.boolean({ char: 'k', description: 'keep previous profiles sent by sshpm' }),
   conf: flags.string({ char: 'c', description: 'application configuration file' }),
   exec: flags.string({ char: 'x', description: 'path to executable required by exporter' })