+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- -i flag to create non-existent configuration files
+
 ### Changed
 
 - Error if registry entries do not exist